Complete lasagna recipe page
Added link to home page to top of recipe page. Added image to show
finished product. Added description and paragraphs to tell some
backstory about how I remember this recipe. Of course, the story was
completely false and I found this recipe on AllRecipes, but this is
just for fun anyway! Added ingredient list, instructions, and nutrition
information. Supplied each section with just a little bit of salty
language to keep things interesting.

Created images directory and refreshed to confirm that relative link
works. No issues with page navigation or function as of this commit.